Update UX and refactor code
Update the name of package javaFX -> javafx to match the package name
format.
Update GUI: app icon, app name, textbox colors.
Add javadoc comments to multiple files.
Fix the bug in returning help messages.
Rename duke -> barry in all the src files.
