--- src/main/java/barry/javaFX/DialogBox.java
@@ -1,74 +1,0 @@
-package barry.javaFX;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.HBox;
-
-/**
- * Represents a dialog box consisting of an ImageView to represent the speaker's face
- * and a label containing text from the speaker.
- */
-public class DialogBox extends HBox {
-    @FXML
-    private Label dialog;
-    @FXML
-    private ImageView displayPicture;
-
-    private DialogBox(String text, Image img) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-            fxmlLoader.setController(this);
-            fxmlLoader.setRoot(this);
-            fxmlLoader.load();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        dialog.setText(text);
-        displayPicture.setImage(img);
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        setAlignment(Pos.TOP_LEFT);
-        dialog.getStyleClass().add("reply-label");
-    }
-
-    // after loading FXML & creating the label
-    public static DialogBox getUserDialog(String text, Image iv) {
-        DialogBox db = new DialogBox(text, iv);
-        db.dialog.getStyleClass().add("user-label");
-        db.setAlignment(Pos.TOP_RIGHT);
-        return db;
-    }
-
-    public static DialogBox getDukeDialog(String text, Image iv) {
-        DialogBox db = new DialogBox(text, iv);
-        db.dialog.getStyleClass().add("reply-label");
-        db.setAlignment(Pos.TOP_LEFT);
-        db.flip();
-        return db;
-    }
-
-    public static DialogBox getErrorDialog(String text, Image iv) {
-        DialogBox db = new DialogBox(text, iv);
-        db.dialog.getStyleClass().add("error-label");
-        db.setAlignment(Pos.TOP_LEFT);
-        return db;
-    }
-}
