+++ src/main/java/barry/javafx/Main.java
@@ -1,4 +1,4 @@
-package barry.javaFX;
+package barry.javafx;
 
 import java.io.IOException;
 
@@ -6,6 +6,7 @@ import barry.Barry;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
+import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
@@ -24,6 +25,10 @@ public class Main extends Application {
             Scene scene = new Scene(ap);
             stage.setScene(scene);
             fxmlLoader.<MainWindow>getController().setBarry(barry); // inject the Source.Barry instance
+            stage.setTitle("Barry — Personal Task Assistant"); // <- shows on the OS title bar
+            stage.getIcons().add(new Image(
+                    getClass().getResource("/images/DaBarry.png").toExternalForm() // optional app icon
+            ));
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
